Introduces LoggerWrapper (to prepare for implementation of Error hook) + cleans up imports to eliminate circular reference with index.ts
