+++ samples/deno-sandbox/import_map.json
@@ -6,9 +6,9 @@
     "../../src/AutoPollConfigService": "../../src/AutoPollConfigService.ts",
     "../../src/Cache": "../../src/Cache.ts",
     "../../src/ConfigCatClient": "../../src/ConfigCatClient.ts",
-    "../../src/ConfigCatClientCache": "../../src/ConfigCatClientCache.ts",
     "../../src/ConfigCatClientOptions": "../../src/ConfigCatClientOptions.ts",
     "../../src/ConfigCatLogger": "../../src/ConfigCatLogger.ts",
+    "../../src/ConfigFetcher": "../../src/ConfigFetcher.ts",
     "../../src/ConfigServiceBase": "../../src/ConfigServiceBase.ts",
     "../../src/FlagOverrides": "../../src/FlagOverrides.ts",
     "../../src/LazyLoadConfigService": "../../src/LazyLoadConfigService.ts",
