+++ src/ConfigFetcher.ts
@@ -1,0 +1,36 @@
+import { OptionsBase } from "./ConfigCatClientOptions";
+
+export enum FetchStatus {
+    Fetched = 0,
+    NotModified = 1,
+    Errored = 2,
+}
+
+export class FetchResult {
+    public status: FetchStatus;
+    public responseBody: string;
+    public eTag?: string;
+
+    constructor(status: FetchStatus, responseBody: string, eTag?: string) {
+        this.status = status;
+        this.responseBody = responseBody;
+        this.eTag = eTag
+    }
+
+    static success(responseBody: string, eTag: string): FetchResult {
+        return new FetchResult(FetchStatus.Fetched, responseBody, eTag);
+    }
+
+    static notModified(): FetchResult {
+        return new FetchResult(FetchStatus.NotModified, "");
+    }
+
+    static error(): FetchResult {
+        return new FetchResult(FetchStatus.Errored, "");
+    }
+
+}
+
+export interface IConfigFetcher {
+    fetchLogic(options: OptionsBase, lastEtag: string | null, callback: (result: FetchResult) => void): void;
+}
