+++ src/ManualPollConfigService.ts
@@ -1,6 +1,6 @@
-import { IConfigService, ConfigServiceBase } from "./ConfigServiceBase";
 import { ManualPollOptions } from "./ConfigCatClientOptions";
-import { IConfigFetcher } from "./index";
+import type { IConfigFetcher } from "./ConfigFetcher";
+import { ConfigServiceBase, IConfigService } from "./ConfigServiceBase";
 import { ProjectConfig } from "./ProjectConfig";
 
 export class ManualPollConfigService extends ConfigServiceBase<ManualPollOptions> implements IConfigService {
