+++ test/ConfigServiceBaseTests.ts
@@ -1,12 +1,13 @@
-import "mocha";
-import { OptionsBase, AutoPollOptions, LazyLoadOptions, ManualPollOptions } from "../src/ConfigCatClientOptions";
-import { FetchResult, ICache, IConfigFetcher, ProjectConfig } from "../src";
-import { AutoPollConfigService } from "../src/AutoPollConfigService";
-import { Mock, It, Times, EqualMatchingInjectorConfig, ResolvedPromiseFactory, RejectedPromiseFactory } from 'moq.ts';
-import { ReturnsAsyncPresetFactory, ThrowsAsyncPresetFactory, MimicsResolvedAsyncPresetFactory, MimicsRejectedAsyncPresetFactory, RootMockProvider, Presets } from "moq.ts/internal";
-import { LazyLoadConfigService } from "../src/LazyLoadConfigService";
 import { assert } from "chai";
-import { InMemoryCache } from "../src/Cache";
+import "mocha";
+import { EqualMatchingInjectorConfig, It, Mock, RejectedPromiseFactory, ResolvedPromiseFactory, Times } from 'moq.ts';
+import { MimicsRejectedAsyncPresetFactory, MimicsResolvedAsyncPresetFactory, Presets, ReturnsAsyncPresetFactory, RootMockProvider, ThrowsAsyncPresetFactory } from "moq.ts/internal";
+import { AutoPollConfigService } from "../src/AutoPollConfigService";
+import { ICache, InMemoryCache } from "../src/Cache";
+import { AutoPollOptions, LazyLoadOptions, ManualPollOptions, OptionsBase } from "../src/ConfigCatClientOptions";
+import { FetchResult, IConfigFetcher } from "../src/ConfigFetcher";
+import { LazyLoadConfigService } from "../src/LazyLoadConfigService";
+import { ProjectConfig } from "../src/ProjectConfig";
 import { delay } from "../src/Utils";
 import { FakeCache } from "./helpers/fakes";
 import { ManualPollConfigService } from "../src/ManualPollConfigService";
