+++ test/helpers/fakes.ts
@@ -1,4 +1,9 @@
-import { FetchResult, ICache, IConfigCatKernel, IConfigCatLogger, IConfigFetcher, LogLevel, OptionsBase, ProjectConfig } from "../../src";
+import { IConfigCatKernel } from "../../src";
+import { ICache } from "../../src/Cache";
+import { OptionsBase } from "../../src/ConfigCatClientOptions";
+import { IConfigCatLogger, LogLevel } from "../../src/ConfigCatLogger";
+import { FetchResult, IConfigFetcher } from "../../src/ConfigFetcher";
+import { ProjectConfig } from "../../src/ProjectConfig";
 
 export class FakeLogger implements IConfigCatLogger {
     public messages: [LogLevel, string][] = [];
@@ -7,36 +12,24 @@ export class FakeLogger implements IConfigCatLogger {
 
     reset() { this.messages.splice(0); }
 
-    isLogLevelEnabled(logLevel: LogLevel): boolean {
-        return this.level >= logLevel;
-    }
-
     log(message: string): void {
         this.info(message);
     }
 
     debug(message: string): void {
-        if (this.isLogLevelEnabled(LogLevel.Debug)) {
-            this.messages.push([LogLevel.Debug, message]);
-        }
+        this.messages.push([LogLevel.Debug, message]);
     }
 
     info(message: string): void {
-        if (this.isLogLevelEnabled(LogLevel.Info)) {
-            this.messages.push([LogLevel.Info, message]);
-        }
+        this.messages.push([LogLevel.Info, message]);
     }
 
     warn(message: string): void {
-        if (this.isLogLevelEnabled(LogLevel.Warn)) {
-            this.messages.push([LogLevel.Warn, message]);
-        }
+        this.messages.push([LogLevel.Warn, message]);
     }
 
     error(message: string): void {
-        if (this.isLogLevelEnabled(LogLevel.Error)) {
-            this.messages.push([LogLevel.Error, message]);
-        }
+        this.messages.push([LogLevel.Error, message]);
     }
 }
